requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audiocraft==1.1.0
audioread==3.0.1
av==11.0.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
clip==0.2.0
cloudpickle==3.0.0
clu==0.0.12
cmake==3.27.7
colorlog==6.7.0
confection==0.1.3
contextlib2==21.6.0
contourpy==1.2.0
crcmod==1.7
cycler==0.12.1
cymem==2.0.8
datasets==2.16.1
decorator==4.4.2
demucs==4.0.1
dill==0.3.7
dmvr==0.0.1
dnspython==2.5.0
docker-pycreds==0.4.0
docopt==0.6.2
dora_search==0.1.12
einops==0.7.0
encodec==0.1.1
etils==1.5.2
exceptiongroup==1.1.3
fastapi==0.104.1
fastavro==1.9.3
fasteners==0.19
ffmpy==0.3.1
filelock==3.13.1
flashy==0.0.2
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
hdfs==2.7.3
hear21passt==0.0.18
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.17.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.2
joblib==1.3.2
Js2Py==0.74
jsonpickle==3.0.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
julius==0.2.7
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
laion-clap==1.1.2
lameenc==1.6.3
langcodes==3.3.0
lazy_loader==0.3
libclang==18.1.1
librosa==0.10.1
lightning-utilities==0.9.0
lit==17.0.5
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
num2words==0.5.13
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objsize==0.6.1
omegaconf==2.3.0
opencv-python==4.8.1.78
openunmix==1.2.1
orjson==3.9.10
packaging==23.2
pandas==2.1.3
pathy==0.10.3
Pillow==10.1.0
pip==23.3
platformdirs==4.0.0
pooch==1.8.0
preshed==3.0.9
proglog==0.1.10
progressbar==2.5
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.8
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydot==1.4.2
pydub==0.25.1
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
pyre-extensions==0.0.23
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
retrying==1.3.4
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scenedetect==0.6.2
scikit-learn==1.3.2
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==68.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
submitit==1.5.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thinc==8.1.12
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.14.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchlibrosa==0.1.0
torchmetrics==1.2.0
torchvision==0.15.1
tqdm==4.66.1
transformers==4.35.1
treetable==0.2.5
triton==2.0.0
typer==0.7.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.10
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
wheel==0.41.2
xformers==0.0.17
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
zstandard==0.22.0
